+++ src/test/java/christmas/ChristmasUserTest.java
@@ -1,7 +1,8 @@
-package model.user;
+package christmas;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import christmas.Application;
+import model.user.ChristmasUser;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
