+++ src/test/java/christmas/VisitDateTest.java
@@ -1,7 +1,8 @@
-package exception;
+package christmas;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import christmas.Application;
+import exception.VisitDate;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
